Revert "refactor: splitted queries.ts into files in folder /queries/"
This reverts commit a7211750329032bbb010d42a1034b3f1a1942727.
