--- packages/data-layer/src/queries/applications/getApplicationStatusByRoundIdAndCID.ts
@@ -1,20 +1,0 @@
-import { gql } from "graphql-request";
-
-export const getApplicationStatusByRoundIdAndCID = gql`
-  query getApplicationStatusByRoundIdAndCID(
-    $roundId: String!
-    $chainId: Int!
-    $metadataCid: String!
-  ) {
-    applications(
-      first: 1
-      filter: {
-        roundId: { equalTo: $roundId }
-        chainId: { equalTo: $chainId }
-        metadataCid: { equalTo: $metadataCid }
-      }
-    ) {
-      status
-    }
-  }
-`;
