--- packages/data-layer/src/queries/applications/index.ts
@@ -1,7 +1,0 @@
-export { getApplicationsByProjectIds } from "./getApplicationsByProjectIds";
-export { getApplicationsByRoundIdAndProjectIds } from "./getApplicationsByRoundIdAndProjectIds";
-export { getApplicationsForExplorer } from "./getApplicationsForExplorer";
-export { getApplicationsForManager } from "./getApplicationsForManager";
-export { getApplicationStatusByRoundIdAndCID } from "./getApplicationStatusByRoundIdAndCID";
-export { getApprovedApplication } from "./getApprovedApplication";
-export { getApprovedApplicationsByProjectIds } from "./getApprovedApplicationsByProjectIds";
