--- packages/data-layer/src/queries/donations/index.ts
@@ -1,2 +1,0 @@
-export { getDirectDonationsByProjectId } from "./getDirectDonationsByProjectId";
-export { getDonationsByDonorAddress } from "./getDonationsByDonorAddress";
