--- packages/data-layer/src/queries/index.ts
@@ -1,8 +1,0 @@
-export * from "./applications";
-export * from "./donations";
-export * from "./programs";
-export * from "./projects";
-export * from "./rounds";
-
-export { getBlockNumberQuery } from "./getBlockNumberQuery";
-export { getPayoutsByChainIdRoundIdProjectId } from "./getPayoutsByChainIdRoundIdProjectId";
