--- packages/data-layer/src/queries/programs/index.ts
@@ -1,2 +1,0 @@
-export { getProgramById } from "./getProgramById";
-export { getProgramsByUserAndTag } from "./getProgramsByUserAndTag";
