--- packages/data-layer/src/queries/projects/getProjects.ts
@@ -1,39 +1,0 @@
-import { gql } from "graphql-request";
-
-/**
- * Get projects by their address
- * @param $alloVersion - The version of Allo
- * @param $first - The number of projects to return
- * @param $chainId - The network ID of the chain
- *
- * @returns The project[]
- */
-export const getProjects = gql`
-  query getProjectsQuery(
-    $alloVersion: [String!]!
-    $first: Int!
-    $chainId: Int!
-  ) {
-    projects(
-      filter: {
-        tags: { equalTo: $alloVersion }
-        not: { tags: { contains: "program" } }
-      }
-      first: $first
-      condition: { chainId: $chainId }
-    ) {
-      id
-      chainId
-      metadata
-      metadataCid
-      name
-      nodeId
-      projectNumber
-      registryAddress
-      tags
-      nonce
-      anchorAddress
-      projectType
-    }
-  }
-`;
