--- packages/data-layer/src/queries/projects/index.ts
@@ -1,8 +1,0 @@
-export { getLegacyProjectId } from "./getLegacyProjectId";
-export { getPaginatedProjects } from "./getPaginatedProjects";
-export { getProjectAnchorByIdAndChainId } from "./getProjectAnchorByIdAndChainId";
-export { getProjects } from "./getProjects";
-export { getProjectsAndRolesByAddress } from "./getProjectsAndRolesByAddress";
-export { getProjectsByAddress } from "./getProjectsByAddress";
-export { getProjectsById } from "./getProjectsById";
-export { getProjectsBySearchTerm } from "./getProjectsBySearchTerm";
