--- packages/data-layer/src/queries/rounds/generateRoundManagerFieldsForQuery.ts
@@ -1,34 +1,0 @@
-export const generateRoundManagerFieldsForQuery = `
-  id
-  chainId
-  applicationsStartTime
-  applicationsEndTime
-  donationsStartTime
-  donationsEndTime
-  matchTokenAddress
-  roundMetadata
-  roundMetadataCid
-  applicationMetadata
-  applicationMetadataCid
-  strategyAddress
-  strategyName
-  readyForPayoutTransaction
-  projectId
-  matchAmount
-  matchAmountInUsd
-  createdByAddress
-  fundedAmount
-  fundedAmountInUsd
-  matchingDistribution
-  roles(first: 100) {
-    role
-    address
-    createdAtBlock
-  }
-  tags
-  project {
-    id
-    name
-    metadata
-  }
-`;
