--- packages/data-layer/src/queries/rounds/getRoundForManager.ts
@@ -1,17 +1,0 @@
-import { gql } from "graphql-request";
-import { generateRoundManagerFieldsForQuery } from "./generateRoundManagerFieldsForQuery";
-
-export const getRoundForManager = gql`
-  query getRoundForManager($roundId: String!, $chainId: Int!) {
-    rounds(
-      first: 1
-      filter: {
-        id: { equalTo: $roundId },
-        chainId: { equalTo: $chainId },
-        roundMetadata: { isNull: false }
-      }
-    ) {
-      ${generateRoundManagerFieldsForQuery}
-    }
-  }
-`;
