--- packages/data-layer/src/queries/rounds/index.ts
@@ -1,6 +1,0 @@
-export { getRoundByIdAndChainId } from "./getRoundByIdAndChainId";
-export { getRoundForExplorer } from "./getRoundForExplorer";
-export { getRoundForManager } from "./getRoundForManager";
-export { getRoundsForManager } from "./getRoundsForManager";
-export { getRoundsForManagerByAddress } from "./getRoundsForManagerByAddress";
-export { getRoundsQuery } from "./getRoundsQuery";
